Reworked CLI Options
The old cli argument structure allowed multiple actions to be taken, but that meant that some options might apply to multiple commands and had to be considered carefully. Effectively, it felt like was causing more problems than it was solving at this point, so I reworked it into a more familiar `subcommand` model.

This does force multiple invocations to, say, remove, then add, then list, but it makes reasoning about the commands and options much simpler.

Docs still need to be updated.

Signed-off-by: Geoffrey Wiseman <geoffrey.wiseman@codiform.com>